Backend/disease_api/requirements.txt
pandas==2.1.1
scikit-learn==1.3.2
joblib==1.3.2
streamlit==1.26.1
numpy==1.27.0
python-dotenv==1.0.1